+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Use the `Copy` implementation for `Clone` and the `Ord` implementation for
+  `PartialOrd` when custom bounds are present.
+
 ## [1.2.7] - 2023-12-14
 
 ### Fixed