+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2025-04-21
 
 ### Added
 - Allow lifetime `for<'a, ...>` bounds in non-bounded generic parameters.
@@ -112,7 +112,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[unreleased]: https://github.com/ModProg/derive-where/compare/v1.2.7...HEAD
+[unreleased]: https://github.com/ModProg/derive-where/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/ModProg/derive-where/compare/v1.2.7...v1.3.0
 [1.2.7]: https://github.com/ModProg/derive-where/compare/v1.2.6...v1.2.7
 [1.2.6]: https://github.com/ModProg/derive-where/compare/v1.2.5...v1.2.6
 [1.2.5]: https://github.com/ModProg/derive-where/compare/v1.2.4...v1.2.5